--- a/waspc/data/Cli/starters/basic/package.json
+++ b/waspc/data/Cli/starters/basic/package.json
@@ -1,6 +1,7 @@
 {
   "name": "__waspAppName__",
   "type": "module",
+  "workspaces": [".wasp/out/*", ".wasp/build/*"],
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",