--- a/waspc/data/Generator/templates/react-app/package.json
+++ b/waspc/data/Generator/templates/react-app/package.json
@@ -1,6 +1,6 @@
 {{={= =}=}}
 {
-  "name": "{= appName =}",
+  "name": "{= packageName =}",
   "version": "0.0.0",
   "private": true,
   "type": "module",