--- a/waspc/data/Generator/templates/server/package.json
+++ b/waspc/data/Generator/templates/server/package.json
@@ -1,6 +1,6 @@
 {{={= =}=}}
 {
-  "name": "server",
+  "name": "{= packageName =}",
   "version": "0.0.0",
   "private": true,
   "type": "module",