Closes #3130

WIP