@infomiho 2025-09-24
Can we get rid of these warnings when running in dev?
```
[ Client!] npm warn config ignoring workspace config at /Users/ilakovac/dev/wasp/waspc/examples/todoApp/.wasp/out/web-app/.npmrc
[ Server!] npm warn config ignoring workspace config at /Users/ilakovac/dev/wasp/waspc/examples/todoApp/.wasp/out/server/.npmrc
```