@infomiho 2025-09-24
I benchmarked it a little bit with the `basic` template running the `wasp compile` command.

**Using workspaces is around 3 seconds faster and it produces around ~50 MB smaller `node_modules`.**

<details>
<summary>Details</summary>

| Setup             | Disk Usage | Size Reduction          |
|-------------------|------------|-------------------------|
| **With Workspace**  | **359M**   | **~13.3% smaller**        |
| **Without Workspace** | 414M       | -                       |

```
hyperfine --warmup 1 \
  'wasp new benchmark_wasp && cd benchmark_wasp && wasp compile && cd .. && rm -rf benchmark_wasp' \
  'wasp-cli new benchmark_lw && cd benchmark_lw && wasp-cli compile && cd .. && rm -rf benchmark_lw'
Benchmark 1: wasp new benchmark_wasp && cd benchmark_wasp && wasp compile && cd .. && rm -rf benchmark_wasp
  Time (mean ± σ):     20.701 s ±  5.990 s    [User: 19.823 s, System: 10.092 s]
  Range (min … max):   18.536 s … 37.737 s    10 runs
 
  Warning: The first benchmarking run for this command was significantly slower than the rest (37.737 s). This could be caused by (filesystem) caches that were not filled until after the first run. You are already using the '--warmup' option which helps to fill these caches before the actual benchmark. You can either try to increase the warmup count further or re-run this benchmark on a quiet system in case it was a random outlier. Alternatively, consider using the '--prepare' option to clear the caches before each timing run.
 
Benchmark 2: wasp-cli new benchmark_lw && cd benchmark_lw && wasp-cli compile && cd .. && rm -rf benchmark_lw
  Time (mean ± σ):     17.714 s ±  6.238 s    [User: 16.703 s, System: 8.253 s]
  Range (min … max):   14.840 s … 35.005 s    10 runs
 
  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.
 
Summary
  wasp-cli new benchmark_lw && cd benchmark_lw && wasp-cli compile && cd .. && rm -rf benchmark_lw ran
    1.17 ± 0.53 times faster than wasp new benchmark_wasp && cd benchmark_wasp && wasp compile && cd .. && rm -rf benchmark_wasp
  ```
</details>