@infomiho 2025-09-25
I've noticed one thing while testing it on `ask-the-document` and migrating: I have to put the values in the `workspaces` field in the right order.  I found that slightly annoying since usually it really doesn't matter and we are being stricter than we have to.

What do you say we do something like `set(what they have).includes(set(what we need))`? This is a nitpick, but again, it's user facing so I'm thinking it's worth to make it more flexible?