@cprecioso 2025-10-28
@sodic ready for re-review

- Undid the validator refactor and made an smaller change for validation
- When defining the globs, convert from StrongPath to regular String earlier to simplify code
- Turned `PackageJson.workspaces` into a `Set` as it fits better how I'm checking it
- Updated `main` and solved conflicts